Move all code into /src for easier coverage
